+++ R/check_heteroscedasticity.R
@@ -12,6 +12,7 @@
 #' @examples
 #' m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
 #' check_heteroscedasticity(m)
+#' plot(check_heteroscedasticity(m))
 #'
 #' @importFrom stats residuals df.residual fitted anova pchisq
 #' @importFrom insight print_color
@@ -41,6 +42,9 @@ check_heteroscedasticity.default <- function(x, ...) {
   } else {
     insight::print_color(sprintf("OK: Error variance appears to be homoscedastic (p = %.3f).", p.val), "green")
   }
+
+  attr(p.val, "object_name") <- deparse(substitute(x), width.cutoff = 500)
+  class(p.val) <- unique(c("check_heteroscedasticity", "see_check_heteroscedasticity", class(p.val)))
 
   invisible(p.val)
 }