+++ R/check_normality.R
@@ -17,6 +17,7 @@
 #' @examples
 #' m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
 #' check_normality(m)
+#' plot(check_normality(m))
 #'
 #' @importFrom stats shapiro.test rstandard
 #' @export