+++ R/check_distribution.R
@@ -12,7 +12,8 @@
 #'
 #' Choosing the right distributional family for regression models is essential to get more accurate estimates and standard errors. This function may help to check a models' distributional family and see if the model-family probably should be reconsidered. Since it is difficult to exactly predict the correct model family, consider this function as somewhat experimental.
 #'
-#' @param model A model (that should response to \code{residuals()}).
+#' @param model Typically, a model (that should response to \code{residuals()}).
+#'   May also be a numeric vector.
 #'
 #' @note This function is somewhat experimental and might be improved in future releases.
 #'   The final decision on the model-family should also be based on theoretical
@@ -46,56 +47,49 @@
 #' @importFrom insight get_response
 #' @export
 check_distribution <- function(model) {
+  UseMethod("check_distribution")
+}
 
+
+#' @export
+check_distribution.numeric <- function(model) {
+  if (!requireNamespace("randomForest", quietly = TRUE)) {
+    stop("Package `randomForest` required for this function to work. Please install it.", call. = FALSE)
+  }
+
+  dat <- .extract_features(model)
+  dist <- as.data.frame(t(stats::predict(classify_distribution, dat, type = "prob")))
+
+  out <- data.frame(
+    Distribution = rownames(dist),
+    p_Vector = dist[[1]],
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+
+  class(out) <- unique(c("check_distribution_numeric", "see_check_distribution_numeric", class(out)))
+  attr(out, "data") <- model
+
+  out
+}
+
+
+#' @export
+check_distribution.default <- function(model) {
   if (!requireNamespace("randomForest", quietly = TRUE)) {
     stop("Package `randomForest` required for this function to work. Please install it.", call. = FALSE)
   }
 
   x <- stats::residuals(model)
   # x_scaled <- .normalize(x)
-
-  # Extract features
-  dat <- data.frame(
-    "SD" = stats::sd(x),
-    "MAD" = stats::mad(x, constant = 1),
-    "Mean_Median_Distance" = mean(x) - median(x),
-    "Mean_Mode_Distance" = mean(x) - as.numeric(bayestestR::map_estimate(x, bw = "nrd0")),
-    "SD_MAD_Distance" = stats::sd(x) - stats::mad(x, constant = 1),
-    "Var_Mean_Distance" = stats::var(x) - mean(x),
-    "Range_SD" = diff(range(x)) / stats::sd(x),
-    "Range" = diff(range(x)),
-    "IQR" = stats::IQR(x),
-    "Skewness" = .skewness(x),
-    "Kurtosis" = .kurtosis(x),
-    "Uniques" = length(unique(x)) / length(x),
-    "N_Uniques" = length(unique(x)),
-    "Min" = min(x),
-    "Max" = max(x)
-  )
+  dat <- .extract_features(x)
 
   dist_residuals <- as.data.frame(t(stats::predict(classify_distribution, dat, type = "prob")))
 
 
-  x <- .factor_to_numeric(insight::get_response(model))
-
   # Extract features
-  dat <- data.frame(
-    "SD" = stats::sd(x),
-    "MAD" = stats::mad(x, constant = 1),
-    "Mean_Median_Distance" = mean(x) - median(x),
-    "Mean_Mode_Distance" = mean(x) - as.numeric(bayestestR::map_estimate(x, bw = "nrd0")),
-    "SD_MAD_Distance" = stats::sd(x) - stats::mad(x, constant = 1),
-    "Var_Mean_Distance" = stats::var(x) - mean(x),
-    "Range_SD" = diff(range(x)) / stats::sd(x),
-    "Range" = diff(range(x)),
-    "IQR" = stats::IQR(x),
-    "Skewness" = .skewness(x),
-    "Kurtosis" = .kurtosis(x),
-    "Uniques" = length(unique(x)) / length(x),
-    "N_Uniques" = length(unique(x)),
-    "Min" = min(x),
-    "Max" = max(x)
-  )
+  x <- .factor_to_numeric(insight::get_response(model))
+  dat <- .extract_features(x)
 
   dist_response <- as.data.frame(t(stats::predict(classify_distribution, dat, type = "prob")))
 
@@ -111,4 +105,25 @@ check_distribution <- function(model) {
   attr(out, "object_name") <- deparse(substitute(model), width.cutoff = 500)
 
   out
+}
+
+
+.extract_features <- function(x) {
+  data.frame(
+    "SD" = stats::sd(x),
+    "MAD" = stats::mad(x, constant = 1),
+    "Mean_Median_Distance" = mean(x) - median(x),
+    "Mean_Mode_Distance" = mean(x) - as.numeric(bayestestR::map_estimate(x, bw = "nrd0")),
+    "SD_MAD_Distance" = stats::sd(x) - stats::mad(x, constant = 1),
+    "Var_Mean_Distance" = stats::var(x) - mean(x),
+    "Range_SD" = diff(range(x)) / stats::sd(x),
+    "Range" = diff(range(x)),
+    "IQR" = stats::IQR(x),
+    "Skewness" = .skewness(x),
+    "Kurtosis" = .kurtosis(x),
+    "Uniques" = length(unique(x)) / length(x),
+    "N_Uniques" = length(unique(x)),
+    "Min" = min(x),
+    "Max" = max(x)
+  )
 }