+++ NEWS.md
@@ -10,6 +10,7 @@
 * `check_model()` now performs more checks for GLM(M)s and other model objects.
 * `r2_nakagawa()` now returns r-squared for models with singular fit, where no random effect variances could be computed. The r-squared then does not take random effect variances into account. This behaviour was changed to be in line with `MuMIn::r.squaredGLMM()`, which returned a value for models with singular fit.
 * `check_distribution()` now detects negative binomial and zero-inflated distributions. Furthermore, attempt to improve accuracy.
+* `check_distribution()` now also accepts a numeric vector as input.
 
 ## New check-functions