+++ NEWS.md
@@ -1,4 +1,4 @@
-# performance 0.2.1
+# performance 0.3.0
 
 ## General